+++ platforms/hermes-javascript/.eslintrc.json
@@ -14,6 +14,7 @@
         "no-console": "warn",
         "quotes": ["error", "single", { "avoidEscape": true }],
         "quote-props": ["error", "as-needed"],
-        "typescript/no-unused-vars": "off"
+        "typescript/no-unused-vars": "off",
+        "no-unused-vars": "warn"
     }
 }
