+++ platforms/hermes-javascript/src/api/types/HermesOptions.ts
@@ -1,0 +1,13 @@
+export type HermesOptions = {
+    address?: string,
+    logs?: boolean,
+    libraryPath?: string,
+    username?: string,
+    password?: string,
+    tls_hostname?: string,
+    tls_ca_file?: string[],
+    tls_ca_path?: string[],
+    tls_client_key?: string,
+    tls_client_cert?: string,
+    tls_disable_root_store?: string
+}
