+++ platforms/hermes-javascript/src/api/types/index.ts
@@ -1,0 +1,3 @@
+export * from './HermesOptions'
+export * from './DialogFlow'
+export * from './ApiSubset'
